% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_join.R
\name{spatial_join}
\alias{spatial_join}
\title{Join attributes from one feature to another based on the spatial relationship.}
\usage{
spatial_join(x, y, xeast, xnorth, yeast, ynorth)
}
\arguments{
\item{x}{Join dataframe.}

\item{y}{Target dataframe.}

\item{xeast}{Easting in the Join dataframe.}

\item{xnorth}{Northing in the Join dataframe.}

\item{yeast}{Easting in the Target dataframe.}

\item{ynorth}{Northing in the Target dataframe.}
}
\value{
The spatial join and attributes from dataframes \code{x} and \code{y}.
}
\description{
Join attributes from one feature to another based on the spatial relationship.
}

